The favicon doesn't match the style of the rest of the page On the form, donation, and posts page the right side section should have its scroll-y set to auto not scroll. Having it set at scroll will force the scroll bar on windows machines even if the div doesn't need the space.
| Form
the like button on posts going almost black would be better if it simply changed to liked or something similar
When I made a post it didn't like the picture I tried to upload, it didn't tell me anything went wrong or what was wrong with my image. you might want error handling for that.

| Form -> replies
its a bit jarring that the size of replies depends on the size of the largest reply.
the reply modal makes it hard to write a long(er) message, you can't comfortably write a message up to the reply max length.

| Donations
Not bad but the styling for the log could be improved a little bit.
If there is no logs it doesn't show a message when on the all section. there should be a message there too.

| Posts
It should auto-select the first post you have, so that the page isn't empty when you open it.
the styling breaks a little bit if a post has a lot of text, you could prob fix this with a max-width.